setup a simple wrapper around the twitter client config
I'll probably change this because it feels weird to create a class that
will only be used for returning an instance of another class, but I'll
look up the appropriate way to do that later. the important part here is
the api keys are kept out of the repo and they're extraction is
abstracted.