first pass on a rough script to find _ebooks accounts
the main heuristic i'm using here is that if an account ends in _ebooks,
then I don't need to see it on my timeline. right now there's no
database, just a text file. the script does a search for users
satisfying this condition and adds them to the list if they're not
already on it.

unfortunately twitter's search results are by 'relevance' which means
that it will be a long time before the list of accounts becomes very
thorough. their API allows for 180 search pings every 15 minutes so this
script will be a perfect candidate to run in a cron job.

TODO: add a whitelist to make sure that non-ebooks accounts using this
format are able to be permanently removed from the block list.