new block script to block collected accounts
before blocking everyone in the 'database' it first removes any accounts
you are already blocking or following to reduce the request size to
twitter's api. this won't scale once the list gets large enough but you
can run the script more than once if you get a rate limit error
(obviously this needs improvement).

it also prints out a list of anyone that is following you that is about
to be blocked, but it should wait for input before actually blocking
them. ideally you would be able to whitelist certain accounts so they
would never be blocked.

the other problem here is that it uses the use context established by
the client which is hard-coded in the configuration. as long you fork
the code and input your own keys then you can block people for different
accounts, but we can do better.