reorganize the BlockList account to EbookAccount
BlockList was designed as a wrapper around accessing the 'database', so
it felt weird to initialize it and then use instance variables and other
such calls. I've changed it's name and purpose to somewhat more like an
ORM wrapper, albeit a very simople one where the only queries are for
every record.