wrap up dependencies and path configration in one place
I'm not really sure this is a good or correct way to do this, as I
rarely ever write straight up ruby scripts outside of a framework. My
approach was to wrap up object configuration or creation
(specifically for the client and logger) in a module that also tracks
the paths so they can all be referenced from one place.

including this module certainly gets the job done but I feel weird
including a module into the global namespace in each script...