a super rudimentary ebooks account model
since I basically only ever used this class as a fake ORM, I'm not sure
if I need to add any other methods to it. It will be nice to update
screen names if they change by correlating the twitter ids, but I'll
worry about that when it might be an actual issue.