actually block ebooks!
by POSTing to EbookAccountsController#block you can now block all of the
users on the main blocklist. the current approach is BAD because it
doesn't take any parameters and auto-creates a new list of accounts to
be blocked. this makes the block_confirmation page lie, since new
accounts could be added or removed when the actual blocking occurs.

to fix this I'll need to submit parameters to the actual block action,
but I'm feeling too sick right to do that :(