a rake task to find more ebooks accounts
this is translated pretty directly from the original script - search
all pages for _ebooks users, filter them out, remove ones that are
already in the db, save the new ones.

I've separated all of the direct calls to the Twitter::REST::Client
behind my own TwitterClient object. When I get around to adding error
handling or backgrounding for these calls hopefully that will make it
easier to do those in one place.

the whenever gem is a nice way to integrate cronjobs into the app. I've
set this new task to run ever 15 minutes since that's the reset period
on twitter's rate limiting, but I won't need to ping that often to find
new accounts.