make a seperate white list file that's easy to read
the primary points of db/account_list.txt and db/global_whitelist.txt are
to provide a quick file-view of what accounts are known of and for
seeding the database on a new fork. they also provide an easy place to
make pull requests to add new accounts to either.

because of this last functionality, db/seeds.rb has been updated so that
the databased can be re-seeded at any point and not create a ton of
duplicates, and also to update the whitelist status of accounts that
already in it.